+++ src/components/TestToolsModal.tsx
@@ -1,4 +1,4 @@
-import React, {useMemo} from 'react';
+import React from 'react';
 import {useOnyx} from 'react-native-onyx';
 import useIsAuthenticated from '@hooks/useIsAuthenticated';
 import useLocalize from '@hooks/useLocalize';
@@ -6,9 +6,8 @@ import useResponsiveLayout from '@hooks/useResponsiveLayout';
 import useStyleUtils from '@hooks/useStyleUtils';
 import useThemeStyles from '@hooks/useThemeStyles';
 import useWindowDimensions from '@hooks/useWindowDimensions';
-import {getBrowser, isChromeIOS} from '@libs/Browser';
 import Navigation from '@navigation/Navigation';
-import toggleTestToolsModal from '@userActions/TestTool';
+import toggleTestToolsModal, {shouldShowProfileTool} from '@userActions/TestTool';
 import CONST from '@src/CONST';
 import ONYXKEYS from '@src/ONYXKEYS';
 import ROUTES from '@src/ROUTES';
@@ -39,17 +38,6 @@ function TestToolsModal() {
     const isAuthenticated = useIsAuthenticated();
     const route = getRouteBasedOnAuthStatus(isAuthenticated, activeRoute);
 
-    const shouldShowProfileTool = useMemo(() => {
-        const browser = getBrowser();
-        const isSafariOrFirefox = browser === CONST.BROWSER.SAFARI || browser === CONST.BROWSER.FIREFOX;
-
-        if (isSafariOrFirefox || isChromeIOS()) {
-            return false;
-        }
-
-        return true;
-    }, []);
-
     return (
         <Modal
             isVisible={!!isTestToolsModalOpen}
@@ -67,7 +55,7 @@ function TestToolsModal() {
                 >
                     {translate('initialSettingsPage.troubleshoot.releaseOptions')}
                 </Text>
-                {shouldShowProfileTool && <ProfilingToolMenu />}
+                {shouldShowProfileTool() && <ProfilingToolMenu />}
                 <ClientSideLoggingToolMenu />
                 {!!shouldStoreLogs && (
                     <TestToolRow title={translate('initialSettingsPage.troubleshoot.debugConsole')}>
