+++ src/libs/actions/TestTool.ts
@@ -1,5 +1,6 @@
 import throttle from 'lodash/throttle';
 import Onyx from 'react-native-onyx';
+import {getBrowser, isChromeIOS} from '@libs/Browser';
 import CONST from '@src/CONST';
 import ONYXKEYS from '@src/ONYXKEYS';
 import * as Modal from './Modal';
@@ -29,4 +30,15 @@ function toggleTestToolsModal() {
     throttledToggle();
 }
 
+function shouldShowProfileTool() {
+    const browser = getBrowser();
+    const isSafariOrFirefox = browser === CONST.BROWSER.SAFARI || browser === CONST.BROWSER.FIREFOX;
+
+    if (isSafariOrFirefox || isChromeIOS()) {
+        return false;
+    }
+    return true;
+}
+
+export {shouldShowProfileTool};
 export default toggleTestToolsModal;
