+++ src/pages/settings/Troubleshoot/TroubleshootPage.tsx
@@ -27,12 +27,11 @@ import useResponsiveLayout from '@hooks/useResponsiveLayout';
 import useThemeStyles from '@hooks/useThemeStyles';
 import useWaitForNavigation from '@hooks/useWaitForNavigation';
 import {setShouldMaskOnyxState} from '@libs/actions/MaskOnyx';
-import {getBrowser, isChromeIOS} from '@libs/Browser';
 import ExportOnyxState from '@libs/ExportOnyxState';
 import Navigation from '@libs/Navigation/Navigation';
 import {clearOnyxAndResetApp} from '@userActions/App';
 import {navigateToConciergeChat} from '@userActions/Report';
-import CONST from '@src/CONST';
+import {shouldShowProfileTool} from '@userActions/TestTool';
 import type {TranslationPaths} from '@src/languages/types';
 import ONYXKEYS from '@src/ONYXKEYS';
 import ROUTES from '@src/ROUTES';
@@ -63,17 +62,6 @@ function TroubleshootPage() {
             ExportOnyxState.shareAsFile(JSON.stringify(dataToShare));
         });
     }, [shouldMaskOnyxState]);
-
-    const shouldShowProfileTool = useMemo(() => {
-        const browser = getBrowser();
-        const isSafariOrFirefox = browser === CONST.BROWSER.SAFARI || browser === CONST.BROWSER.FIREFOX;
-
-        if (isSafariOrFirefox || isChromeIOS()) {
-            return false;
-        }
-
-        return true;
-    }, []);
 
     const menuItems = useMemo(() => {
         const debugConsoleItem: BaseMenuItem = {
@@ -150,7 +138,7 @@ function TroubleshootPage() {
                     >
                         <View style={[styles.flex1, styles.mt5]}>
                             <View>
-                                {shouldShowProfileTool && <ProfilingToolMenu />}
+                                {shouldShowProfileTool() && <ProfilingToolMenu />}
                                 <ClientSideLoggingToolMenu />
                                 <TestToolRow title={translate('initialSettingsPage.troubleshoot.maskExportOnyxStateData')}>
                                     <Switch
